[refactor] Stamp 도메인 관련 클래스 리펙토링 주석 삭제 response dto 스트림 메서드명 : from<엔티티명>List -> fromEntities EncryptService 디렉토리 변경 : /stamp/service > /global/utils
