+++ src/main/java/com/halo/eventer/domain/stamp/controller/MissionController.java
@@ -14,16 +14,15 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/stamp/mission")
 public class MissionController {
+
   private final MissionService missionService;
 
-  /** 미션 단일 조회 */
   @MissionDetailGetApi
   @GetMapping
   public MissionDetailGetDto getMission(@RequestParam("missionId") Long missionId) {
     return missionService.getMission(missionId);
   }
 
-  /** 미션 단일 수정 */
   @MissionUpdateApi
   @PatchMapping
   public void updateMission(
